Air Control and Monitoring Device
It is a device used to check the quality of air, temperature and humidity in a room and would publish the data on Thingspeak cloud. It also includes sending of E-mail telling the basic precautions for the air quality and humidity of the room using Twilio cloud. It was done on Texas Instrumentation's CC3200 Launchpad with various sensors like MQ135 and DHT11.